collectively cleaned up files and organised file structure, rearraged styling.
