+++ src/App.js
@@ -1,13 +1,10 @@
 import './App.css';
-import Header from './Header.js';
-import hr from './assets/hr.png'
-
-// import chatBubble from './assets/Speech-Bubble.png'
-import UserInfo from './UserInfo';
-import PokemonSelections from './PokemonSelections';
-import Results from './Results';
-// import InvestigationAlgorithm from './InvestigationAlgorithm';
-import Footer from './Footer.js';
+import Header from './components/Header'
+import Introduction from './components/Introduction';
+import UserInfo from './components/UserInfo';
+import PokemonSelections from './components/PokemonSelections';
+import Results from './components/Results';
+import Footer from './components/Footer.js';
 import { useState } from 'react';
 
 function App() {
@@ -22,13 +19,7 @@ function App() {
     <>
       <div className = 'wrapper'>
         <Header />
-        <div className = 'introduction'>
-          <img className='bubble' src={hr} alt='HR person introduces detective for Pokemon Detective game'></img>
-          <div className = 'openingText'>
-            <p>Hello there! Welcome to the Scotland Yard’s Pokémon Investigation Unit. This special unit solves crimes with the help of Pokémon. These Pokémon are specially trained and qualified to assist PIU detectives with their cases.</p>
-            <p>Oh! I see you don’t have your PIU ID with you – it must be your first day as a Detective! Congratulations! Let me help you get your ID set up. Can you give me your name?</p>        
-          </div>
-        </div>
+        <Introduction />
         <UserInfo
           updateUserName={setUserName}
           currentUserName={userName}
@@ -40,14 +31,10 @@ function App() {
           ? <PokemonSelections selectedCrime={chosenCrimeType} setPType={setPType} />
           : null
         }
-        {
-          pType !== ""
-          ? <Results 
+        <Results 
           crimeType={chosenCrimeType}
           pokiType={pType}
-          />
-          : null
-        }
+        />
       </div>
         <Footer />
     </>
