+++ src/components/Footer.js
@@ -4,7 +4,7 @@ function Footer() {
     return(
     <footer>
         <div className="wrapper">
-            <p>Created at <a target="_blank" rel="noopener noreferrer" href="https://junocollege.com/">Juno College of Technology</a> using the <a target="_blank" rel="noopener noreferrer" href="https://pokeapi.co/docs/v2">Pokemon</a> and <a target="_blank" rel="noopener noreferrer" href="https://data.police.uk/docs/">UK Police</a> APIs. Character art by <a href="https://openclipart.org/detail/311169/talk-show-host">qubodup</a>, and cork background by <a href="https://unsplash.com/photos/KsAo8ouBn8A">Josephine Bredehoft</a>.</p>
+            <p>Created at <a target="_blank" rel="noopener noreferrer" href="https://junocollege.com/">Juno College of Technology</a> using the <a target="_blank" rel="noopener noreferrer" href="https://pokeapi.co/docs/v2">Pokémon</a> and <a target="_blank" rel="noopener noreferrer" href="https://data.police.uk/docs/">UK Police</a> APIs. Character art by <a href="https://openclipart.org/detail/311169/talk-show-host">qubodup</a>, and cork background by <a href="https://unsplash.com/photos/KsAo8ouBn8A">Josephine Bredehoft</a>.</p>
             
             <p className="portfolioLinks">Created by: <a href="https://www.jennifercollins.ca/" target="_blank" rel="noopener noreferrer">Jennifer Collins</a>, <a href="https://kimberleycodes.com/" target="_blank" rel="noopener noreferrer">Kimberley Schubert</a>, <a href="https://marinaz.dev/" target="_blank" rel="noopener noreferrer">Marina Zubtsova</a>, <a href="https://maxinepower.ca/" target="_blank" rel="noopener noreferrer">Maxine Power</a>.</p>
         </div>
