+++ src/components/Header.js
@@ -3,7 +3,7 @@
 function Header() {
     return(
         <header>
-            <h1>Detective Pokemon</h1>
+            <h1>Detective Pokémon</h1>
         </header>
     );
 }
