+++ src/components/PokemonCard.js
@@ -1,12 +1,7 @@
 // PokemonCard
-
-import { useState } from "react";
-
-
 function PokemonCard(props) {
-   
+    
     const selectedPokemonType = (event) => {
-        const pokemonType = props.type
         props.setPType(event.target.value)
     }
